requirements/core.txt


aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.0.4
filelock==3.12.0
frozenlist==1.3.3
fsspec[http]==2023.5.0
huggingface-hub==0.15.1
idna==3.4
joblib==1.2.0
lightning-utilities==0.8.0
multidict==6.0.4
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pillow==9.5.0
pytorch-lightning==1.9.5
pyyaml==6.0
regex==2023.5.5
requests==2.31.0
safetensors==0.3.1
timm==0.9.2
#torch==1.13.1+cpu
#torchmetrics==0.11.4
torchmetrics
#torchvision==0.14.1+cpu
tqdm==4.65.0
typing-extensions==4.6.2
urllib3==2.0.2
yarl==1.9.2
